maint: remove classes marked with deprecated_class and functions marked with derprecate decorator; move definitions to a new maintenance module
