+++ yt/__init__.py
@@ -80,9 +80,6 @@ from yt.units import (
 from yt.units.unit_object import define_unit
 from yt.utilities.logger import set_log_level, ytLogger as mylog
 
-# For backwards compatibility
-TimeSeriesData = deprecated_class(DatasetSeries)
-
 from yt.frontends.api import _frontend_container
 
 frontends = _frontend_container()
@@ -92,9 +89,7 @@ from yt.frontends.ytdata.api import save_as_dataset
 
 # For backwards compatibility
 GadgetDataset = frontends.gadget.GadgetDataset
-GadgetStaticOutput = deprecated_class(GadgetDataset)
 TipsyDataset = frontends.tipsy.TipsyDataset
-TipsyStaticOutput = deprecated_class(TipsyDataset)
 
 import yt.visualization.volume_rendering.api as volume_rendering
 from yt.loaders import simulation  # deprecated alias for load_simulation
@@ -166,7 +161,7 @@ from yt.visualization.volume_rendering.api import (
 
 def _check_deprecated_parameters():
     from yt.config import ytcfg
-    from yt.funcs import issue_deprecation_warning
+    from yt.maintenance.deprecation import issue_deprecation_warning
 
     if ytcfg.getboolean("yt", "loadfieldplugins"):
         issue_deprecation_warning(
