+++ yt/data_objects/index_subobjects/octree_subset.py
@@ -8,8 +8,8 @@ import yt.geometry.particle_smooth as particle_smooth
 from yt.data_objects.selection_objects.data_selection_objects import (
     YTSelectionContainer,
 )
-from yt.funcs import issue_deprecation_warning, mylog
 from yt.geometry.particle_oct_container import ParticleOctreeContainer
+from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.units.dimensions import length
 from yt.units.yt_array import YTArray
 from yt.utilities.exceptions import (
@@ -18,6 +18,7 @@ from yt.utilities.exceptions import (
     YTParticleDepositionNotImplemented,
 )
 from yt.utilities.lib.geometry_utils import compute_morton
+from yt.utilities.logger import ytLogger as mylog
 
 
 def cell_count_cache(func):
