+++ yt/fields/particle_fields.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from yt.fields.derived_field import ValidateParameter, ValidateSpatial
-from yt.funcs import issue_deprecation_warning
+from yt.maintenance.deprecation import issue_demeshening_deprecation_warning
 from yt.units.yt_array import uconcatenate, ucross
 from yt.utilities.lib.misc_utilities import (
     obtain_position_vector,
@@ -848,17 +848,6 @@ def add_particle_average(
     return fn
 
 
-DEP_MSG_SMOOTH_FIELD = (
-    "Since yt-4.0, it's no longer necessary to add a field specifically for "
-    "smoothing, because the global octree is removed. The old behavior of "
-    "interpolating onto a grid structure can be recovered through data objects "
-    "like ds.arbitrary_grid, ds.covering_grid, and most closely ds.octree. The "
-    "visualization machinery now treats SPH fields properly by smoothing onto "
-    "pixel locations. See this page to learn more: "
-    "https://yt-project.org/doc/yt4differences.html"
-)
-
-
 def add_volume_weighted_smoothed_field(
     ptype,
     coord_name,
@@ -870,7 +859,7 @@ def add_volume_weighted_smoothed_field(
     nneighbors=64,
     kernel_name="cubic",
 ):
-    issue_deprecation_warning("This function is deprecated. " + DEP_MSG_SMOOTH_FIELD)
+    issue_demeshening_deprecation_warning("This function is deprecated. ")
 
 
 def add_nearest_neighbor_field(ptype, coord_name, registry, nneighbors=64):
