+++ yt/fields/species_fields.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from yt.fields.field_detector import FieldDetector
 from yt.frontends.sph.data_structures import ParticleDataset
-from yt.funcs import issue_deprecation_warning
+from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.utilities.chemical_formulas import ChemicalFormula
 from yt.utilities.physical_ratios import _primordial_mass_fraction
