+++ yt/frontends/fits/data_structures.py
@@ -12,9 +12,10 @@ from more_itertools import always_iterable
 from yt.config import ytcfg
 from yt.data_objects.index_subobjects.grid_patch import AMRGridPatch
 from yt.data_objects.static_output import Dataset
-from yt.funcs import issue_deprecation_warning, mylog, setdefaultattr
+from yt.funcs import mylog, setdefaultattr
 from yt.geometry.geometry_handler import YTDataChunk
 from yt.geometry.grid_geometry_handler import GridIndex
+from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.units import dimensions
 from yt.units.unit_lookup_table import default_unit_symbol_lut, unit_prefixes
 from yt.units.unit_object import UnitParseError
@@ -737,17 +738,10 @@ class SpectralCubeFITSDataset(SkyDataFITSDataset):
         parameters=None,
         units_override=None,
         unit_system="cgs",
-        z_axis_decomp=None,
     ):
         if auxiliary_files is None:
             auxiliary_files = []
         self.spectral_factor = spectral_factor
-        if z_axis_decomp is not None:
-            issue_deprecation_warning(
-                "The 'z_axis_decomp' argument is deprecated, "
-                "as this decomposition is now performed for "
-                "spectral-cube FITS datasets automatically."
-            )
         super().__init__(
             filename,
             nprocs=nprocs,
