+++ yt/utilities/answer_testing/framework.py
@@ -24,8 +24,9 @@ from nose.plugins import Plugin
 from yt.config import ytcfg
 from yt.data_objects.static_output import Dataset
 from yt.data_objects.time_series import SimulationTimeSeries
-from yt.funcs import get_pbar, issue_deprecation_warning
+from yt.funcs import get_pbar
 from yt.loaders import load, load_simulation
+from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.testing import (
     assert_allclose_units,
     assert_almost_equal,
