+++ yt/utilities/particle_generator.py
@@ -1,6 +1,7 @@
 import numpy as np
 
-from yt.funcs import get_pbar, issue_deprecation_warning
+from yt.funcs import get_pbar
+from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.units.yt_array import uconcatenate
 from yt.utilities.lib.particle_mesh_operations import CICSample_3
