+++ yt/visualization/eps_writer.py
@@ -3,7 +3,7 @@ import pyx
 from matplotlib import cm, pyplot as plt
 
 from yt.config import ytcfg
-from yt.funcs import issue_deprecation_warning
+from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.units.unit_object import Unit
 from yt.units.yt_array import YTQuantity
 from yt.utilities.logger import ytLogger as mylog
