+++ yt/visualization/image_writer.py
@@ -3,7 +3,8 @@ import builtins
 import numpy as np
 
 from yt.config import ytcfg
-from yt.funcs import get_brewer_cmap, get_image_suffix, issue_deprecation_warning, mylog
+from yt.funcs import get_brewer_cmap, get_image_suffix, mylog
+from yt.maintenance.deprecation import issue_deprecation_warning
 from yt.units.yt_array import YTQuantity
 from yt.utilities import png_writer as pw
 from yt.utilities.exceptions import YTNotInsideNotebook
